created payment form and fixed call Server
